Entity does not "auto" update any more. Register explicit updater to renderer in Intialize()/DoInitialize() method. This allows external animators to be attached to the object and can also attach multiple updaters to the objet (e.g. local animation plus additional path animator, for example).

+++ src/renderer.h
@@ -15,7 +15,7 @@
 #include <boost/function.hpp>
 #include <boost/unordered_map.hpp>
 
-typedef boost::function<bool(long)> UpdateFunction;
+typedef boost::function<bool(float)> UpdateFunction;
 
 class Renderer : public Worker
 {
